requirements.txt
Django==3.1.3
django-environ==0.4.5
psycopg2==2.8.6
graphene-django==2.13.0
graphene-federation==0.1.0
coverage==5.3.1
factory-boy==3.2.0
ipython==7.19.0
django-extensions==3.1.0
pyjwt==1.7.1
django-graphql-jwt==0.3.0
django-graphql-auth==0.3.15
